Added backend route token authorization

+++ backend/controllers/games-controllers.js
@@ -26,6 +26,11 @@ module.exports.getPublicList = async (req, res, next) => {
 module.exports.getUserList = async (req, res, next) => {
   const userId = req.params.uid;
   
+  if (parseInt(userId) !== req.userData.userId) {
+    const error = new HttpError('Permission to access resource denied.', 401);
+    return next(error);
+  }
+
   let userGames;
   try {
     userGames = await GameRepo.findByUser(userId);
@@ -47,7 +52,7 @@ module.exports.getUserList = async (req, res, next) => {
 
 module.exports.postSaveGame = async (req, res, next) => {
   const errors = validationResult(req);
-
+  
   if (!errors.isEmpty()) {
     const error = new HttpError('Invalid inputs, please check your data.', 422);
     return next(error);
@@ -55,6 +60,11 @@ module.exports.postSaveGame = async (req, res, next) => {
 
   const { gameObject } = req.body;
   
+  if (gameObject.userId !== req.userData.userId) {
+    const error = new HttpError('Permission to save resource denied.', 401);
+    return next(error);
+  }
+
   let game;
   try {
     game = await GameRepo.insert(gameObject);
@@ -101,6 +111,11 @@ module.exports.patchEditGame = async (req, res, next) => {
   const gameId = req.params.gid;
   const { gameObject } = req.body;
 
+  if (gameObject.userId !== req.userData.userId) {
+    const error = new HttpError('Permission to edit resource denied.', 401);
+    return next(error);
+  }
+
   let game;
   try {
     game = await GameRepo.update(gameId, gameObject);
@@ -121,6 +136,19 @@ module.exports.deleteGame = async (req, res, next) => {
   const gameId = req.params.gid;
 
   let game;
+
+  try {
+    game = await GameRepo.findById(gameId);
+  } catch (err) {
+    const error = new HttpError('Something went wrong, please try again.', 500);
+    return next(error);
+  }
+
+  if (game.user_id !== req.userData.userId) {
+    const error = new HttpError('Permission to delete resource denied.', 401);
+    return next(error);
+  }
+
   try {
     game = await GameRepo.delete(gameId);
   } catch (err) {
